refactor(src):입력 진행 함수 유효성 검사후 Retry하는 기능 리펙토링
입력 진행 함수 유효성  검사후 Retry하는 기능 리펙토링
