+++ docs/README.md
@@ -2,9 +2,15 @@
 
 ## 클래스별 기능 목록
 
+### BridgeController클래스
+
+- [x] 다리 길이 입력값 받아오는 기능
+- [x] 다리 이동 입력값 받아오는 기능
+- [x] 재시작/종료 입력값 받아오는 기능
+
 ### BridgeMaker클래스
 
-- [ ] 다리 길이 생성 기능
+- [x] 다리 길이 생성 기능
 
 ## 입력
