+++ src/validation/commandValidation.js
@@ -1,3 +1,5 @@
+const catchError = require('../utils/catchError');
+
 function commandValidation(commandInput) {
   if (commandInput !== 'Q' && commandInput !== 'R') {
     return false;
@@ -6,4 +8,13 @@ function commandValidation(commandInput) {
   return true;
 }
 
-module.exports = commandValidation;
+function checkRetryCommand(input) {
+  if (!commandValidation(input)) {
+    catchError('[ERROR] 재시작,종료 R혹은 Q인 문자여야 합니다.');
+    return false;
+  }
+
+  return true;
+}
+
+module.exports = { commandValidation, checkRetryCommand };
