+++ src/validation/movingValidation.js
@@ -1,3 +1,5 @@
+const catchError = require('../utils/catchError');
+
 function movingValidation(movingInput) {
   if (movingInput !== 'U' && movingInput !== 'D') {
     return false;
@@ -6,4 +8,13 @@ function movingValidation(movingInput) {
   return true;
 }
 
-module.exports = movingValidation;
+function checkRetryMoving(input) {
+  if (!movingValidation(input)) {
+    catchError('[ERROR] 다리 이동은 U혹은 D인 문자여야 합니다.');
+    return false;
+  }
+
+  return true;
+}
+
+module.exports = { movingValidation, checkRetryMoving };
